# src/main/resources/openapi/swagger-config.yaml
...
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
  "/api/projects/{project_id}/task-states/{task-states_id}/tasks/create":
    post:
      tags:
        - task-controller
      summary: Create Task for Task State
      description: Create new Task
      operationId: createTask
      parameters:
        - name: project_id
          in: path
          description: project id
          required: true
          schema:
            type: integer
        - name: task-states_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: empty JSON taskDTO
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TaskDTO"
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TaskDTO"
        '400':
          description: Bad request
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TaskDTO"
  "/api/projects/{project_id}/task-states/create":
    post:
      tags:
        - task-state-controller
      summary: Create Task State for project by project id and task state id
      description: Returns new Task State
      operationId: createTaskState
      parameters:
        - name: project_id
          in: path
          description: project id
          required: true
          schema:
            type: integer
      requestBody:
        description: empty JSON taskStateDTO
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TaskStateDTO"
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TaskStateDTO"
        '400':
          description: Bad request
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TaskStateDTO"
...
  "/api/projects/{project_id}/task-states/{task-states_id}/tasks/update/{task_id}":
    patch:
      tags:
        - task-controller
      summary: Update Task
      description: Returns updated Task
      operationId: updateTaskById
      parameters:
        - name: project_id
          in: path
          description: project id
          required: true
          schema:
            type: integer
        - name: task_id
          in: path
          description: task id
          required: true
          schema:
            type: integer
        - name: task-states_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TaskDTO"
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TaskDTO"
        '400':
          description: Bad request
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TaskDTO"
        '404':
          description: Not found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TaskDTO"
  "/api/projects/{project_id}/task-states/update/{task-state_id}":
    patch:
      tags:
        - task-state-controller
      summary: Update Task State for project by project id and task state id
      description: Returns updated Task State with Tasks. Updates all fields, but
        if you want to update order, you should change only previous elementFor adding
        new Task see TaskController
      operationId: updateProject
      parameters:
        - name: project_id
          in: path
          description: project id
          required: true
          schema:
            type: integer
        - name: task-state_id
          in: path
          description: task state id
          required: true
          schema:
            type: integer
      requestBody:
        description: JSON taskStateDTO
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TaskStateDTO"
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TaskStateDTO"
        '400':
          description: Bad request
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TaskStateDTO"
  "/api/projects/update/{id}":
    patch:
      tags:
        - Projects
      summary: Update project by id
      description: |
        Returns updated project with Task States. Updates only name.
        For adding new Task States see TaskStateController.
      operationId: updateProject_1
      parameters:
        - name: id
          in: path
          description: project`s id
          required: true
          schema:
            type: integer
      requestBody:
        description: empty JSON projectDTO
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ProjectDTO"
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ProjectDTO"
        '400':
          description: Bad request
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ProjectDTO"
...
  "/api/projects/{project_id}/task-states/{task-states_id}/tasks/get":
    get:
      tags:
        - task-controller
      summary: Get Tasks for Task State by project id and task state id
      description: Returns all Tasks
      operationId: getTasks
      parameters:
        - name: project_id
          in: path
          description: project id
          required: true
          schema:
            type: integer
        - name: task-states_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/TaskDTO"
  "/api/projects/{project_id}/task-states/{task-states_id}/tasks/get/{task_id}":
    get:
      tags:
        - task-controller
      summary: Get Tasks for Task State by project id, task state id and task id
      description: Returns all Tasks
      operationId: getTaskById
      parameters:
        - name: project_id
          in: path
          description: project id
          required: true
          schema:
            type: integer
        - name: task_id
          in: path
          description: task id
          required: true
          schema:
            type: integer
        - name: task-states_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TaskDTO"
        '400':
          description: Bad request
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TaskDTO"
        '404':
          description: Not found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TaskDTO"
  "/api/projects/{project_id}/task-states/get":
    get:
      tags:
        - task-state-controller
      summary: Get all Task States for project by project id
      description: Returns all Task States for project
      operationId: getAllTaskStates
      parameters:
        - name: project_id
          in: path
          description: project id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/TaskStateDTO"
  "/api/projects/{project_id}/task-states/get/{task-state_id}":
    get:
      tags:
        - task-state-controller
      summary: Get Task State for project by project id and task state id
      description: Returns Task State for project
      operationId: getTaskState
      parameters:
        - name: project_id
          in: path
          description: project id
          required: true
          schema:
            type: integer
        - name: task-state_id
          in: path
          description: task state id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TaskStateDTO"
        '404':
          description: Not found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TaskStateDTO"
...
  "/api/projects/{project_id}/task-states/{task-states_id}/tasks/delete/{task_id}":
    delete:
      tags:
        - task-controller
      summary: Delete Task
      description: Delete Task
      operationId: deleteTaskById
      parameters:
        - name: project_id
          in: path
          description: project id
          required: true
          schema:
            type: integer
        - name: task_id
          in: path
          description: task id
          required: true
          schema:
            type: integer
        - name: task-states_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            "*/*":
              schema:
                type: string
                enum:
                  - 100 CONTINUE
                  - 101 SWITCHING_PROTOCOLS
                  - 102 PROCESSING
                  - 103 EARLY_HINTS
                  - 103 CHECKPOINT
                  - 200 OK
                  - 201 CREATED
                  - 202 ACCEPTED
                  - 203 NON_AUTHORITATIVE_INFORMATION
                  - 204 NO_CONTENT
                  - 205 RESET_CONTENT
                  - 206 PARTIAL_CONTENT
                  - 207 MULTI_STATUS
                  - 208 ALREADY_REPORTED
                  - 226 IM_USED
                  - 300 MULTIPLE_CHOICES
                  - 301 MOVED_PERMANENTLY
                  - 302 FOUND
                  - 302 MOVED_TEMPORARILY
                  - 303 SEE_OTHER
                  - 304 NOT_MODIFIED
                  - 305 USE_PROXY
                  - 307 TEMPORARY_REDIRECT
                  - 308 PERMANENT_REDIRECT
                  - 400 BAD_REQUEST
                  - 401 UNAUTHORIZED
                  - 402 PAYMENT_REQUIRED
                  - 403 FORBIDDEN
                  - 404 NOT_FOUND
                  - 405 METHOD_NOT_ALLOWED
                  - 406 NOT_ACCEPTABLE
                  - 407 PROXY_AUTHENTICATION_REQUIRED
                  - 408 REQUEST_TIMEOUT
                  - 409 CONFLICT
                  - 410 GONE
                  - 411 LENGTH_REQUIRED
                  - 412 PRECONDITION_FAILED
                  - 413 PAYLOAD_TOO_LARGE
                  - 413 REQUEST_ENTITY_TOO_LARGE
                  - 414 URI_TOO_integer
                  - 414 REQUEST_URI_TOO_integer
                  - 415 UNSUPPORTED_MEDIA_TYPE
                  - 416 REQUESTED_RANGE_NOT_SATISFIABLE
                  - 417 EXPECTATION_FAILED
                  - 418 I_AM_A_TEAPOT
                  - 419 INSUFFICIENT_SPACE_ON_RESOURCE
                  - 420 METHOD_FAILURE
                  - 421 DESTINATION_LOCKED
                  - 422 UNPROCESSABLE_ENTITY
                  - 423 LOCKED
                  - 424 FAILED_DEPENDENCY
                  - 425 TOO_EARLY
                  - 426 UPGRADE_REQUIRED
                  - 428 PRECONDITION_REQUIRED
                  - 429 TOO_MANY_REQUESTS
                  - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
                  - 451 UNAVAILABLE_FOR_LEGAL_REASONS
                  - 500 INTERNAL_SERVER_ERROR
                  - 501 NOT_IMPLEMENTED
                  - 502 BAD_GATEWAY
                  - 503 SERVICE_UNAVAILABLE
                  - 504 GATEWAY_TIMEOUT
                  - 505 HTTP_VERSION_NOT_SUPPORTED
                  - 506 VARIANT_ALSO_NEGOTIATES
                  - 507 INSUFFICIENT_STORAGE
                  - 508 LOOP_DETECTED
                  - 509 BANDWIDTH_LIMIT_EXCEEDED
                  - 510 NOT_EXTENDED
                  - 511 NETWORK_AUTHENTICATION_REQUIRED
        '400':
          description: Bad request
          content:
            "*/*":
              schema:
                type: string
                enum:
                  - 100 CONTINUE
                  - 101 SWITCHING_PROTOCOLS
                  - 102 PROCESSING
                  - 103 EARLY_HINTS
                  - 103 CHECKPOINT
                  - 200 OK
                  - 201 CREATED
                  - 202 ACCEPTED
                  - 203 NON_AUTHORITATIVE_INFORMATION
                  - 204 NO_CONTENT
                  - 205 RESET_CONTENT
                  - 206 PARTIAL_CONTENT
                  - 207 MULTI_STATUS
                  - 208 ALREADY_REPORTED
                  - 226 IM_USED
                  - 300 MULTIPLE_CHOICES
                  - 301 MOVED_PERMANENTLY
                  - 302 FOUND
                  - 302 MOVED_TEMPORARILY
                  - 303 SEE_OTHER
                  - 304 NOT_MODIFIED
                  - 305 USE_PROXY
                  - 307 TEMPORARY_REDIRECT
                  - 308 PERMANENT_REDIRECT
                  - 400 BAD_REQUEST
                  - 401 UNAUTHORIZED
                  - 402 PAYMENT_REQUIRED
                  - 403 FORBIDDEN
                  - 404 NOT_FOUND
                  - 405 METHOD_NOT_ALLOWED
                  - 406 NOT_ACCEPTABLE
                  - 407 PROXY_AUTHENTICATION_REQUIRED
                  - 408 REQUEST_TIMEOUT
                  - 409 CONFLICT
                  - 410 GONE
                  - 411 LENGTH_REQUIRED
                  - 412 PRECONDITION_FAILED
                  - 413 PAYLOAD_TOO_LARGE
                  - 413 REQUEST_ENTITY_TOO_LARGE
                  - 414 URI_TOO_integer
                  - 414 REQUEST_URI_TOO_integer
                  - 415 UNSUPPORTED_MEDIA_TYPE
                  - 416 REQUESTED_RANGE_NOT_SATISFIABLE
                  - 417 EXPECTATION_FAILED
                  - 418 I_AM_A_TEAPOT
                  - 419 INSUFFICIENT_SPACE_ON_RESOURCE
                  - 420 METHOD_FAILURE
                  - 421 DESTINATION_LOCKED
                  - 422 UNPROCESSABLE_ENTITY
                  - 423 LOCKED
                  - 424 FAILED_DEPENDENCY
                  - 425 TOO_EARLY
                  - 426 UPGRADE_REQUIRED
                  - 428 PRECONDITION_REQUIRED
                  - 429 TOO_MANY_REQUESTS
                  - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
                  - 451 UNAVAILABLE_FOR_LEGAL_REASONS
                  - 500 INTERNAL_SERVER_ERROR
                  - 501 NOT_IMPLEMENTED
                  - 502 BAD_GATEWAY
                  - 503 SERVICE_UNAVAILABLE
                  - 504 GATEWAY_TIMEOUT
                  - 505 HTTP_VERSION_NOT_SUPPORTED
                  - 506 VARIANT_ALSO_NEGOTIATES
                  - 507 INSUFFICIENT_STORAGE
                  - 508 LOOP_DETECTED
                  - 509 BANDWIDTH_LIMIT_EXCEEDED
                  - 510 NOT_EXTENDED
                  - 511 NETWORK_AUTHENTICATION_REQUIRED
        '404':
          description: Not found
          content:
            "*/*":
              schema:
                type: string
                enum:
                  - 100 CONTINUE
                  - 101 SWITCHING_PROTOCOLS
                  - 102 PROCESSING
                  - 103 EARLY_HINTS
                  - 103 CHECKPOINT
                  - 200 OK
                  - 201 CREATED
                  - 202 ACCEPTED
                  - 203 NON_AUTHORITATIVE_INFORMATION
                  - 204 NO_CONTENT
                  - 205 RESET_CONTENT
                  - 206 PARTIAL_CONTENT
                  - 207 MULTI_STATUS
                  - 208 ALREADY_REPORTED
                  - 226 IM_USED
                  - 300 MULTIPLE_CHOICES
                  - 301 MOVED_PERMANENTLY
                  - 302 FOUND
                  - 302 MOVED_TEMPORARILY
                  - 303 SEE_OTHER
                  - 304 NOT_MODIFIED
                  - 305 USE_PROXY
                  - 307 TEMPORARY_REDIRECT
                  - 308 PERMANENT_REDIRECT
                  - 400 BAD_REQUEST
                  - 401 UNAUTHORIZED
                  - 402 PAYMENT_REQUIRED
                  - 403 FORBIDDEN
                  - 404 NOT_FOUND
                  - 405 METHOD_NOT_ALLOWED
                  - 406 NOT_ACCEPTABLE
                  - 407 PROXY_AUTHENTICATION_REQUIRED
                  - 408 REQUEST_TIMEOUT
                  - 409 CONFLICT
                  - 410 GONE
                  - 411 LENGTH_REQUIRED
                  - 412 PRECONDITION_FAILED
                  - 413 PAYLOAD_TOO_LARGE
                  - 413 REQUEST_ENTITY_TOO_LARGE
                  - 414 URI_TOO_integer
                  - 414 REQUEST_URI_TOO_integer
                  - 415 UNSUPPORTED_MEDIA_TYPE
                  - 416 REQUESTED_RANGE_NOT_SATISFIABLE
                  - 417 EXPECTATION_FAILED
                  - 418 I_AM_A_TEAPOT
                  - 419 INSUFFICIENT_SPACE_ON_RESOURCE
                  - 420 METHOD_FAILURE
                  - 421 DESTINATION_LOCKED
                  - 422 UNPROCESSABLE_ENTITY
                  - 423 LOCKED
                  - 424 FAILED_DEPENDENCY
                  - 425 TOO_EARLY
                  - 426 UPGRADE_REQUIRED
                  - 428 PRECONDITION_REQUIRED
                  - 429 TOO_MANY_REQUESTS
                  - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
                  - 451 UNAVAILABLE_FOR_LEGAL_REASONS
                  - 500 INTERNAL_SERVER_ERROR
                  - 501 NOT_IMPLEMENTED
                  - 502 BAD_GATEWAY
                  - 503 SERVICE_UNAVAILABLE
                  - 504 GATEWAY_TIMEOUT
                  - 505 HTTP_VERSION_NOT_SUPPORTED
                  - 506 VARIANT_ALSO_NEGOTIATES
                  - 507 INSUFFICIENT_STORAGE
                  - 508 LOOP_DETECTED
                  - 509 BANDWIDTH_LIMIT_EXCEEDED
                  - 510 NOT_EXTENDED
                  - 511 NETWORK_AUTHENTICATION_REQUIRED
...
components:
  schemas:
    TaskDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 1
        name:
          type: string
          example: This is the name of task
        description:
          type: string
          example: This is the description of task
        created_at:
          type: string
          format: date-time
          readOnly: true
        task_state_id:
          type: integer
          format: int64
          readOnly: true
          example: 1
      example:
        - id: 27
          name: task1
          description: description1
          created_at: '2023-10-24T17:26:05.902537Z'
          task_state_id: 146
    TaskStateDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 1
        name:
          type: string
          example: This is the name of task state
        previous_task_state_id:
          type: integer
          format: int64
          example: 1
        next_task_state_id:
          type: integer
          format: int64
          readOnly: true
          example: 1
        created_at:
          type: string
          format: date-time
          readOnly: true
        project_id:
          type: array
          readOnly: true
          example:
            - id: 27
              name: task1
              description: description1
              created_at: '2023-10-24T17:26:05.902537Z'
              task_state_id: 146
          items:
            "$ref": "#/components/schemas/TaskDTO"
      example:
        - id: 145
          name: In progress
          previous_task_state_id: 144
          next_task_state_id: 146
          created_at: '2023-10-24T17:16:43.498700Z'
          project_id: 57
          task_entities: []
    ProjectDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 1
        name:
          type: string
          example: This is the name of project
        created_at:
          type: string
          format: date-time
          readOnly: true
        user_id:
          type: integer
          format: int64
          readOnly: true
        task_state_entities:
          type: array
          readOnly: true
          example:
            - id: 145
              name: In progress
              previous_task_state_id: 144
              next_task_state_id: 146
              created_at: '2023-10-24T17:16:43.498700Z'
              project_id: 57
              task_entities: []
          items:
            "$ref": "#/components/schemas/TaskStateDTO"
...